feat: refactoring pipeline with references
- merge pipelineRefs with the actual pipeline module
- update all dependencies related to that change
