+++ server/tests/pipeline/test_references.py
@@ -1,7 +1,11 @@
 import pytest
 
-from weaverbird.pipeline.pipeline import Pipeline, PipelineWithVariables
-from weaverbird.pipeline.references import PipelineWithRefs, ReferenceUnresolved
+from weaverbird.pipeline.pipeline import (
+    Pipeline,
+    PipelineWithRefs,
+    PipelineWithVariables,
+    ReferenceUnresolved,
+)
 from weaverbird.pipeline.steps import (
     AppendStepWithRefs,
     DomainStep,
